Optimized city tab component, login and main pages, place sorting component.
