+++ src/pages/login-page/login-page.tsx
@@ -1,31 +1,8 @@
-import { FormEvent, useRef } from 'react';
 import Header from '../../components/common/header';
-import { Link, useNavigate } from 'react-router-dom';
-import { useAppDispatch } from '../../hooks';
-import { login } from '../../store/action-api';
-import { AppRoute } from '../../components/consts';
+import { Link } from 'react-router-dom';
+import LoginForm from './login-form';
 
 function LoginPage(): JSX.Element {
-  const dispatch = useAppDispatch();
-  const navigate = useNavigate();
-  const loginRef = useRef<HTMLInputElement | null>(null);
-  const passwordRef = useRef<HTMLInputElement | null>(null);
-
-  const submitHanlder = (evt: FormEvent<HTMLFormElement>) =>{
-    evt.preventDefault();
-
-    if (loginRef.current !== null && passwordRef.current !== null){
-      dispatch(login({
-        login: loginRef.current.value,
-        password: passwordRef.current.value
-      })).then((response) => {
-        if (response.meta.requestStatus === 'fulfilled'){
-          navigate(AppRoute.Root);
-        }
-      });
-    }
-  };
-
   return (
     <div className="page page--gray page--login">
       <Header showUserLogin={false} />
@@ -33,36 +10,7 @@ function LoginPage(): JSX.Element {
         <div className="page__login-container container">
           <section className="login">
             <h1 className="login__title">Sign in</h1>
-            <form
-              className="login__form form"
-              action="#"
-              method="post"
-              onSubmit={submitHanlder}
-            >
-              <div className="login__input-wrapper form__input-wrapper">
-                <label className="visually-hidden">E-mail</label>
-                <input
-                  className="login__input form__input"
-                  type="email"
-                  name="email"
-                  placeholder="Email"
-                  ref={loginRef}
-                  required
-                />
-              </div>
-              <div className="login__input-wrapper form__input-wrapper">
-                <label className="visually-hidden">Password</label>
-                <input
-                  className="login__input form__input"
-                  type="password"
-                  name="password"
-                  placeholder="Password"
-                  ref={passwordRef}
-                  required
-                />
-              </div>
-              <button className="login__submit form__submit button" type="submit">Sign in</button>
-            </form>
+            <LoginForm/>
           </section>
           <section className="locations locations--login locations--current">
             <div className="locations__item">
