+++ src/pages/main-page/main-page-content.tsx
@@ -1,20 +1,23 @@
 import Map from '../../components/map/map';
-import PlacesSorting from '../../components/sorting/sorting';
+import { PlacesSorting } from '../../components/sorting/sorting';
 import CITIES_MAP from '../../data/cities';
 import { OfferPreviewType } from '../../types/offer-type';
 import { filterOffers } from '../../utils/utils';
 import { OfferCardsList } from '../../components/offers-list/offers-list';
-import { memo } from 'react';
+import { memo, useCallback, useMemo, useState } from 'react';
 
 type Props = {
     offers: OfferPreviewType[];
     currentCity: string;
-    activeOfferId: string | null;
-    handleActiveOfferChange: (id: string | null) => void;
 }
 
-function MainPageContent({offers, currentCity, activeOfferId, handleActiveOfferChange}: Props): JSX.Element{
-  const filteredOffers = filterOffers(offers, currentCity);
+function MainPageContent({offers, currentCity}: Props): JSX.Element{
+  const filteredOffers = useMemo(()=> filterOffers(offers, currentCity), [offers, currentCity]);
+  const [activeOfferId, setActiveOffer] = useState<string | null>(null);
+  const handleActiveOfferChange = useCallback((id: string | null) => {
+    setActiveOffer(id);
+  }, []);
+
   return(
     <div className="cities">
       <div className="cities__places-container container">
